replace construct_meshes by inline imlementation
This patch changes the algorithm by which the interface topologies
skeleton_mesh and ghost_mesh are constructed. Since the new algorithm is
considerabily more concise, the construct_meshes helper function is no longer
deemed required. Instead, a get_ref helper function is introduced for the
benefit of constructing the ghost mesh.